# server/static/openapi-specification.yaml
openapi: 3.0.0

info:
  title: weg-li Recommender API
  version: 0.1.0

servers:
  - url: https://europe-west3-wegli-296209.cloudfunctions.net/api
  - url: https://api.weg-li.de/v1

tags:
  - name: user
    description: Add, remove or update user data.
  - name: analyze
    description: Analyze incident data in order to provide recommendations for important report data.
  - name: report
    description: Submit a new report.

paths:
  /user:
    post:
      tags:
        - user
      description: Create a new user for the analysis platform.
      responses:
        200:
          description: 'The operation was successful.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    $ref: '#/components/schemas/UserID'
                  access_token:
                    type: string
                    description: The corresponding access token for the newly created user account.
  /user/{user_id}:
    delete:
      tags:
        - user
      description: Deletes the information that is associated to the provided anonymous user id on the analysis platform.
      security:
        - accessTokenAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UserID'
      responses:
        200:
          description: 'The operation was successful.'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        409:
          description: 'An endpoint error occurred.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /analyze/data:
    post:
      tags:
        - analyze
      security:
        - accessTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - time
                - location
              properties:
                user_id:
                  $ref: '#/components/schemas/UserID'
                time:
                  type: integer
                  format: int64
                  description: The unix time of the violation in milliseconds.
                  example: 1605481357079
                location:
                  description: The location of the violation.
                  allOf:
                    - $ref: '#/components/schemas/Location'
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  violation:
                    type: array
                    items:
                      type: string
                    minItems: 1
                    maxItems: 5
                    example:
                      - no_stopping
                      - no_parking
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /analyze/image/upload:
    get:
      description: Returns expiring cloud storage urls. Upload urls only accept PUT requests and expect files in jpeg format.
      tags:
        - analyze
      parameters:
        - name: quantity
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 5
      responses:
        200:
          description: 'The operation was successful.'
          content:
            application/json:
              schema:
                type: object
                required:
                  - token
                  - google_cloud_urls
                properties:
                  token:
                    type: string
                    description: 'A unique image token'
                    example: 82571c4a-7f07-4bdc-acdc-5b2745a00de3
                  google_cloud_urls:
                    type: array
                    items:
                      type: string
                    description: 'The Google Cloud Storage urls to which the image(s) should be uploaded.'
                    example:
                      - https://api.google.com/...
  /analyze/image/{image_token}:
    get:
      description: Returns suggestions of the violating vehicle regarding license plate number, make, model and color based on the provided images ordered by descending confidence.
      tags:
        - analyze
      parameters:
        - name: image_token
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'The operation was successful.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suggestions'
        404:
          description: 'No image(s) could be found for the provided token.'
        409:
          description: 'An endpoint error occurred.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /report:
    post:
      tags:
        - report
      security:
        - accessTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - report
              properties:
                user_id:
                  $ref: '#/components/schemas/UserID'
                report:
                  $ref: '#/components/schemas/Report'
      responses:
        200:
          description: 'The operation was successful.'

components:
  responses:
    UnauthorizedError:
      description: The access token is missing or invalid.
  securitySchemes:
    accessTokenAuth:
      type: http
      scheme: bearer
  schemas:
    UserID:
      type: string
      example: 550e8400-e29b-11d4-a716-446655440000
    Location:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          description: The latitude value of the location's coordinates.
          example: 52.550081
        longitude:
          type: number
          format: double
          description: The longitude value of the location's coordinates.
          example: 13.370763
    Suggestions:
      type: object
      properties:
        suggestions:
          type: object
          properties:
            license_plate_number:
              type: array
              items:
                type: string
              description: Suggestions for the license plate number. Empty list when no license plate number recognized.
              example: ['B-WL 1234', 'KA KK 3455']
            make:
              type: array
              items:
                type: string
              description: Suggestions for the make. Empty list when no make recognized.
              example: ['Mercedes-Benz']
            model:
              type: array
              items:
                type: string
              description: Suggestions for the model. Empty list when no model recognized.
              example: ['A-Class']
            color:
              type: array
              items:
                type: string
              description: Suggestions for the color. Empty list when no color recognized.
              example: ['blue', 'white']
    Report:
      type: object
      properties:
        violation_type:
          type: number
          format: int32
          example: 1
        time:
          type: number
          format: int64
          example: 1606756404
          description: The date and time of the violation as Unix timestamp (in seconds since epoch).
        location:
          $ref: '#/components/schemas/Location'
        image_token:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        description:
          type: string